Move annotations to separate component
